Sampada_S_Gowda/EDUCATIONAL_CONTENT-SUMMARIZER/app/requirement.txt
tensorflow
tensorflow-hub
pytesseract
pillow
pdf2image
pypdf2
python-dotenv
numpy
google-generativeai
